requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# MQTT
aiomqtt==1.2.1
paho-mqtt==2.0.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# TimescaleDB / time-series
# (uses standard PostgreSQL driver, TimescaleDB is a PG extension)

# Redis / queue
redis[asyncio]==5.0.4
rq==1.16.2

# AI
anthropic==0.26.0
openai==1.30.1
httpx==0.27.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Notifications
sendgrid==6.11.0

# Utils
orjson==3.10.3
structlog==24.2.0
tenacity==8.3.0
